Separate out IBC in slice files for weird import reasons
